+++ nylas/client/client.py
@@ -8,11 +8,12 @@ from datetime import datetime, timedelta
 from itertools import chain
 
 import requests
+from requests import HTTPError
 from urlobject import URLObject
 import six
 from six.moves.urllib.parse import urlencode
 from nylas._client_sdk_version import __VERSION__
-from nylas.client.errors import MessageRejectedError
+from nylas.client.errors import MessageRejectedError, NylasApiError
 from nylas.client.restful_model_collection import RestfulModelCollection
 from nylas.client.restful_models import (
     Calendar,
@@ -28,6 +29,7 @@ from nylas.client.restful_models import (
     Folder,
     Label,
     Draft,
+    Component,
 )
 from nylas.client.neural_api_models import Neural
 from nylas.client.scheduler_restful_model_collection import (
@@ -60,8 +62,9 @@ def _validate(response):
         # we will handle it separate and handle a _different_ exception
         # so that users don't think they need to pay.
         raise MessageRejectedError(response)
+    elif response.status_code >= 400:
+        raise NylasApiError(response)
 
-    response.raise_for_status()
     return response
 
 
@@ -426,6 +429,9 @@ class APIClient(json.JSONEncoder):
     def scheduler(self):
         return SchedulerRestfulModelCollection(self)
 
+    def components(self):
+        return RestfulModelCollection(Component, self)
+
     @property
     def neural(self):
         return Neural(self)
@@ -438,7 +444,7 @@ class APIClient(json.JSONEncoder):
         # Is this a request for a resource under the accounts/billing/admin
         # namespace (/a)? If the latter, pass the client_secret
         # instead of the secret_token
-        if api_root == "a":
+        if api_root:
             return self.admin_session
         return self.session
 
@@ -446,11 +452,18 @@ class APIClient(json.JSONEncoder):
         # FIXME @karim: remove this interim code when we've got rid
         # of the old accounts API.
         postfix = "/{}".format(extra) if extra else ""
-        if cls.api_root != "a":
-            url = "{}/{}{}".format(self.api_server, cls.collection_name, postfix)
+        path = "/{}".format(cls.collection_name) if cls.collection_name else ""
+        if not cls.api_root:
+            url = "{server}{path}{postfix}".format(
+                server=self.api_server, path=path, postfix=postfix
+            )
         else:
-            url = "{}/a/{}/{}{}".format(
-                self.api_server, self.client_id, cls.collection_name, postfix
+            url = "{server}/{prefix}/{client_id}{path}{postfix}".format(
+                server=self.api_server,
+                prefix=cls.api_root,
+                client_id=self.client_id,
+                path=path,
+                postfix=postfix,
             )
 
         converted_data = create_request_body(filters, cls.datetime_filter_attrs)
@@ -463,25 +476,33 @@ class APIClient(json.JSONEncoder):
         self, cls, id, extra=None, headers=None, stream=False, path=None, **filters
     ):
         """Get an individual REST resource"""
-        headers = headers or {}
-        headers.update(self.session.headers)
-
         if path is None:
-            path = cls.collection_name
+          path = cls.collection_name
         postfix = "/{}".format(extra) if extra else ""
+        path = "/{}".format(path) if path else ""
         id = "/{}".format(id) if id else ""
-        if cls.api_root != "a":
-            url = "{}/{}{}{}".format(self.api_server, path, id, postfix)
+        if not cls.api_root:
+            url = "{server}{path}{id}{postfix}".format(
+                server=self.api_server, path=path, id=id, postfix=postfix
+            )
         else:
-            url = "{}/a/{}/{}{}{}".format(
-                self.api_server, self.client_id, path, id, postfix
+            url = "{server}/{prefix}/{client_id}{path}{id}{postfix}".format(
+                server=self.api_server,
+                prefix=cls.api_root,
+                client_id=self.client_id,
+                path=path,
+                id=id,
+                postfix=postfix,
             )
 
         converted_data = create_request_body(filters, cls.datetime_filter_attrs)
         url = str(URLObject(url).add_query_params(converted_data.items()))
-        response = self._get_http_session(cls.api_root).get(
-            url, headers=headers, stream=stream
-        )
+
+        session = self._get_http_session(cls.api_root)
+
+        headers = headers or {}
+        headers.update(session.headers)
+        response = session.get(url, headers=headers, stream=stream)
         return _validate(response)
 
     def _get_resource(self, cls, id, **filters):
@@ -498,9 +519,15 @@ class APIClient(json.JSONEncoder):
         return response.content
 
     def _create_resource(self, cls, data, **kwargs):
+        name = "{prefix}{path}".format(
+            prefix="/{}/{}".format(cls.api_root, self.client_id)
+            if cls.api_root
+            else "",
+            path="/{}".format(cls.collection_name) if cls.collection_name else "",
+        )
         url = (
             URLObject(self.api_server)
-            .with_path("/{name}".format(name=cls.collection_name))
+            .with_path("{name}".format(name=name))
             .set_query_params(**kwargs)
         )
 
@@ -511,7 +538,7 @@ class APIClient(json.JSONEncoder):
         else:
             converted_data = create_request_body(data, cls.datetime_attrs)
             headers = {"Content-Type": "application/json"}
-            headers.update(self.session.headers)
+            headers.update(session.headers)
             response = session.post(url, json=converted_data, headers=headers)
 
         result = _validate(response).json()
@@ -520,9 +547,13 @@ class APIClient(json.JSONEncoder):
         return cls.create(self, **result)
 
     def _create_resources(self, cls, data):
-        url = URLObject(self.api_server).with_path(
-            "/{name}".format(name=cls.collection_name)
+        name = "{prefix}{path}".format(
+            prefix="/{}/{}".format(cls.api_root, self.client_id)
+            if cls.api_root
+            else "",
+            path="/{}".format(cls.collection_name) if cls.collection_name else "",
         )
+        url = URLObject(self.api_server).with_path("{name}".format(name=name))
         session = self._get_http_session(cls.api_root)
 
         if cls == File:
@@ -532,16 +563,22 @@ class APIClient(json.JSONEncoder):
                 create_request_body(datum, cls.datetime_attrs) for datum in data
             ]
             headers = {"Content-Type": "application/json"}
-            headers.update(self.session.headers)
+            headers.update(session.headers)
             response = session.post(url, json=converted_data, headers=headers)
 
         results = _validate(response).json()
         return [cls.create(self, **x) for x in results]
 
     def _delete_resource(self, cls, id, data=None, **kwargs):
+        name = "{prefix}{path}".format(
+            prefix="/{}/{}".format(cls.api_root, self.client_id)
+            if cls.api_root
+            else "",
+            path="/{}".format(cls.collection_name) if cls.collection_name else "",
+        )
         url = (
             URLObject(self.api_server)
-            .with_path("/{name}/{id}".format(name=cls.collection_name, id=id))
+            .with_path("{name}/{id}".format(name=name, id=id))
             .set_query_params(**kwargs)
         )
         session = self._get_http_session(cls.api_root)
@@ -553,6 +590,18 @@ class APIClient(json.JSONEncoder):
     def _put_resource(self, cls, id, data, extra=None, path=None, **kwargs):
         if path is None:
             path = cls.collection_name
+        name = "{prefix}{path}".format(
+            prefix="/{}/{}".format(cls.api_root, self.client_id)
+            if cls.api_root
+            else "",
+            path="/{}".format(path) if path else "",
+        )
+        url = (
+            URLObject(self.api_server)
+            .with_path("{name}/{id}".format(name=name, id=id))      
+              
+        
+        
         postfix = "/{}".format(extra) if extra else ""
         url = (
             URLObject(self.api_server)
@@ -577,15 +626,17 @@ class APIClient(json.JSONEncoder):
     def _post_resource(self, cls, id, method_name, data, path=None):
         if path is None:
             path = cls.collection_name
-        if cls.api_root != "a":
+        path = "/{}".format(path) if path else ""
+        if not cls.api_root:
             url_path = "/{name}/{id}/{method}".format(
                 name=path, id=id, method=method_name
             )
         else:
             # Management method.
-            url_path = "/a/{client_id}/{name}/{id}/{method}".format(
+            url_path = "/{prefix}/{client_id}{path}/{id}/{method}".format(
+                prefix=cls.api_root,
                 client_id=self.client_id,
-                name=path,
+                path=path,
                 id=id,
                 method=method_name,
             )